+++ packages/talker/lib/talker.dart
@@ -2,8 +2,11 @@ export 'package:talker_logger/talker_logger.dart';
 
 export 'src/extensions/extensions.dart';
 export 'src/filter.dart';
+export 'src/history.dart';
 export 'src/models/models.dart';
 export 'src/observer.dart';
 export 'src/settings.dart';
 export 'src/talker.dart';
 export 'src/talker_key.dart';
+export 'src/utils/utils.dart';
+export 'src/well_known_titles.dart';